Refactor elements, make ProductTemplate.js to be used for both pages
