+++ src/components/card-detail/RecommendationCard.module.css
@@ -1,5 +1,5 @@
 .card {
-    margin-bottom: 35px;
+    margin-bottom: 40px;
 }
 
 .img {
@@ -9,6 +9,6 @@
 
 .card-name {
     font-family: "Montserrat SemiBold", sans-serif;
-    margin-bottom: 3px;
+    margin: 7px 0;
     font-size: 18px;
 }
