+++ src/pages/product/ProductBidPage.js
@@ -1,5 +1,9 @@
-const ProductBidPage = () => {
+import ProductTemplate from "../../components/card-detail/ProductTemplate";
 
+const ProductBidPage = () => {
+  return (
+      <ProductTemplate />
+  )
 }
 
 export default ProductBidPage
